clean project component and pages
