--- src/App.js
@@ -1,11 +1,0 @@
-import Calculator from './components/Calculator';
-
-function App() {
-  return (
-    <div className="App">
-      <Calculator />
-    </div>
-  );
-}
-
-export default App;
